feat(apiserver): service 阶段提交

+++ apiserver/v1/FileChunks.go
@@ -7,7 +7,7 @@ const TableNameGalloFileChunk = "galloFileChunks"
 // FileChunk 文件片段
 type FileChunk struct {
 	metav1.ObjectMeta `json:"metadata,omitempty"`
-	FileID            uint32 `gorm:"column:fileId;comment:文件ID" json:"fileId"` // 文件ID
+	FileID            uint64 `gorm:"column:fileId;comment:文件ID" json:"fileId"` // 文件ID
 	Data              []byte `gorm:"column:data;comment:分块内容" json:"data"`     // 分块内容
 }
